+++ app/frontend/src/javascript/models/setting.ts
@@ -119,7 +119,20 @@ export enum SettingName {
   PackOnlyForSubscription = 'pack_only_for_subscription',
   OverlappingCategories = 'overlapping_categories',
   ExtendedPricesInSameDay = 'extended_prices_in_same_day',
-  PublicRegistrations = 'public_registrations'
+  PublicRegistrations = 'public_registrations',
+  SocialsFacebook = 'facebook',
+  SocialsTwitter = 'twitter',
+  SocialsViadeo = 'viadeo',
+  SocialsLinkedin = 'linkedin',
+  SocialsInstagram = 'instagram',
+  SocialsYoutube = 'youtube',
+  SocialsVimeo = 'vimeo',
+  SocialsDailymotion = 'dailymotion',
+  SocialsGithub = 'github',
+  SocialsEchosciences = 'echosciences',
+  SocialsPinterest = 'pinterest',
+  SocialsLastfm = 'lastfm',
+  SocialsFlickr = 'flickr'
 }
 
 export type SettingValue = string|boolean|number;