+++ app/frontend/src/javascript/models/social-network.ts
@@ -1,0 +1,20 @@
+export interface SocialNetwork {
+  name: string,
+  url: string
+}
+
+export const supportedNetworks = [
+  'facebook',
+  'twitter',
+  'viadeo',
+  'linkedin',
+  'instagram',
+  'youtube',
+  'vimeo',
+  'dailymotion',
+  'github',
+  'echosciences',
+  'pinterest',
+  'lastfm',
+  'flickr'
+];