+++ app/models/setting.rb
@@ -130,7 +130,20 @@ class Setting < ApplicationRecord
                              extended_prices_in_same_day
                              public_registrations
                              accounting_Pack_code
-                             accounting_Pack_label] }
+                             accounting_Pack_label
+                             facebook
+                             twitter
+                             viadeo
+                             linkedin
+                             instagram
+                             youtube
+                             vimeo
+                             dailymotion
+                             github
+                             echosciences
+                             pinterest
+                             lastfm
+                             flickr] }
   # WARNING: when adding a new key, you may also want to add it in:
   # - config/locales/en.yml#settings
   # - app/frontend/src/javascript/models/setting.ts#SettingName